Reset file structure to correspond with the spec
